example: rename greeting demo
